SSL error handling: Remove unnecessary confusing code
I'm confident this is unnecessary. The original bug in #3283 was
to call ignoreSslErrors() without an argument in the 'accept'
case, which meant ignoring *all* subsequent SSL errors.

With that fixed, explicitly aborting the reply and resetting QNAM
is not needed since not ignoring the error will lead to the SSL
handshake failing.

See also:
  75b38d1a2ffe57d0f1eb3ebb8c5f30b8b2a185e4 (workaround introduced)
  89376e14d6135a6f39a6df99d54fde253573575c (real fix)
  76ce5adbf02af052ab8d7596b8ded75c7d4d7fcb (cherry-pick of workaround)